moved sidebar to its own component
